+++ arm_client/signal/correlation/simpleCorrelator.cpp
@@ -8,7 +8,7 @@ void SimpleCorrelator::correlation(Signal signal, Signal pattern, RealNum* out) 
         for (int j = 0; j < pattern.size(); j++)
         {
             if (i + j < signal.size())
-                out[i] += signal[i + j].toFloat() * pattern[j].toFloat();
+                out[i] += signal[i + j].toInt() * pattern[j].toInt();
         }
     }
     //emit correlationCalculated(corr);
